feat: two-layer speech model — shell owns content, module mounts say
- Speech: base abstraction, __content__ kernel command owned by Shell
- TTSSpeech: extended abstraction, say command mounted by SpeechChannelModule
- Shell always has content (MockSpeech fallback), module no-op without TTSSpeech
- Remove make_content_command_from_speech() and TTSSpeech.commands() from contracts
- Add build_content_command() + SpeechChannelModule(register_content=False)
- SpeechChannelModule: only registers say when isinstance(speech, TTSSpeech)
- speech_channel.py migrates to channel.with_module(SpeechChannelModule(register_content=True))
- StatefulChannel adds with_module() abstract method
- TTSSpeechServiceProvider: singleton=True

coding by deepseek-v4

via claude code
