+++ src/ghoshell_moss/core/speech/__init__.py
@@ -2,6 +2,7 @@ from ghoshell_common.contracts import LoggerItf
 
 from ghoshell_moss.contracts.speech import TTS, Speech, SpeechStream, StreamAudioPlayer
 from ghoshell_moss.core.speech.mock import MockSpeech
+from ghoshell_moss.core.speech.speech_module import SpeechChannelModule, build_content_command
 from ghoshell_moss.core.speech.stream_tts_speech import BaseTTSSpeech, TTSSpeechStream
